almost forgot perms

+++ commands/util/requestx.js
@@ -23,15 +23,6 @@ module.exports = class RequestXCommand extends commando.Command {
 			]
 		});
 	}
-	
-	hasPermission(msg) {
-		if(msg.client.isOwner(msg.author)){
-			return true;
-		}
-		else{
-			return msg.member.permissions.has('MANAGE_MESSAGES');
-		}
-	}
 
 	async run(msg, args) {
 		msg.delete();
@@ -46,7 +37,7 @@ module.exports = class RequestXCommand extends commando.Command {
 		var xLimit = msg.client.provider.get(msg.guild, 'xLimit' + msg.channel.id, 7)
 		var xlogChannelIDs = msg.client.provider.get(msg.guild, 'xlogChannelIDs', []);
 		var xmsg = msg.channel.messages.get(args.id);
-		if(xmsg != null){
+		if(xmsg != null && !msg.client.provider.get(msg.guild, 'xBlacklistIDs', []).includes(msg.author.id)){
 			sendMessages(xlogChannelIDs, `Placed an ❌ on ${xmsg.author.username}[${xmsg.author.id}]'s message[${xmsg.id}] by request of ${msg.author.username}[${msg.author.id}].`)
 			xLimit > 1 ? xmsg.react('\u{274c}') : xmsg.delete();
 		}
